Work on Profil Management
Wrapping the Profil Page Information into the class and sending it on
its merry way will wait until i get the input designed just right to be
practical and pleasant enough.
Deleting and Adding Games is a bit of a hassle though atm, i'll have to
rethink the process. The current one is bad.